// zhkh_frontend/src/app/infrastructures/enums/auditLog.ts
export enum AuditLogEnum {
    EDIT = 'EDIT',
    DELETE = 'DELETE',
    ENTER_TO_SYSTEM = 'ENTER_TO_SYSTEM',
    EXPORT_DATA = 'EXPORT_DATA',
    CREATE = 'CREATE',
    ATTACH = 'ATTACH',
    UNTIE = 'UNTIE',
}

export enum EntityTypeEnum {
    HOUSE = 'HOUSE',
    COMPANY = 'COMPANY',
    USER = 'USER',
    DATA = 'DATA',
}